pyproject.toml
[tool.poetry]
name = "python-django"
version = "0.1.0"
description = ""
authors = ["Aknur Sharipzhan <sharipzhan.aknur@mail.ru>"]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
django = "4.2.11"
djangorestframework = "3.14.0"
django-filter = "^25.1"
django-debug-toolbar = "^6.0.0"
pillow = "^11.3.0"
gunicorn = "^23.0.0"
drf-spectacular = "^0.28.0"
jsonschema = "4.25.1"
jsonschema-specifications = "2025.4.1"
flake8 = "7.3.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
